FIX: Eliminated trivial warnings given by compiler.

+++ OnlineShopping/OnlineShopping/User.h
@@ -55,16 +55,16 @@ void User::remove_from_cart(Item item, int num){
 vector<Order> User::view_history(){
 	vector< vector<string> >  dbOrders = dbManager.get_order_history(name);
 	
-	for(int i=0;i<dbOrders.size();i++){
+	for(unsigned int i=0;i<dbOrders.size();i++){
 		Order current_order;
 		current_order.payment = dbOrders[i][0];
 		current_order.address = dbOrders[i][1];
 		current_order.total = std::stof(dbOrders[i][2]);
-		for(int j=3;i<dbOrders.size();j+=3){
+		for(unsigned int j=3;i<dbOrders.size();j+=3){
 			Item current_item;
 			current_item.name = dbOrders[i][j];
 			current_item.price = std::stof(dbOrders[i][j+2]);
-			current_order.quantity.push_back(std::stof(dbOrders[i][j+2]));
+			current_order.quantity.push_back(std::stoi(dbOrders[i][j+2]));
 			current_order.items.push_back(current_item);
 		}
 		orders.push_back(current_order);
